FlightRecoder
[A test only feature that is not enabled in Bend.]

A feature for recording all text editor surface inputs and playing them back. Anything that is done with the keyboard, mouse on the editor surface area can be captured and played back. However things like commands from the right click menu cannot be captured and are not intended to be captured. Hitting F9 key will save the current rendering surface on to a png file in the current executable directory. During playback F9 will compare against existing image in the FlightRecord.xml directory and write the image to disk only if the images are different.

TextCoreDemo - The test app for the editor surface now has a collection of command line options to enable testing.

/playback=FlightRecord.xml -  play back the previously recorded flight record FlightRecord.xml.
/hud -  enable the debug HUD.
/linenumber -  enable line numbers
/showformatting -  enables show formatting.
/exit -  exit the process after playing back. It will return error code -1, if any images were written to disk.
/hide - makes it so that text code application window is not on screen and tests can run silently.

With this change it is possible to enable end users to submit FlightRecords to diagnose bugs. 
The second cool thing it enables is the ability to run tests without any UI or stealing window focus. 
This should allow for a more extensive test suite and more consitent test validation.

Note: Snapshot format was switched to PNG from JPG because PNG does not have any compression artifacts. 
JPGs were resulting in false failures in the current test framework.

git-svn-id: https://subversion.assembla.com/svn/sushanth/trunk@253 93233f36-2c2b-41c2-b1a5-e0ee276c51f9